refactor(core): extract IListAction family out of list framework
Lift the list-action contract interfaces (IBaseListAction, IListAction,
IOptionalAction, IMultiListAction, IGlobalListAction) out of the legacy
list.component.types into shared/components/signal-list/list-action.types
so they survive deletion of the list framework, and export them from
public-api. Repoint the live consumers (store entity-catalog.types, helm
endpoint generator) + the dead-until-D consumers (list.component,
table-cell-actions) at the new home; base-cf-list-config resolves via the
@stratosui/core barrel. Drop the now-orphaned ActionState import.
