+++ src/frontend/packages/core/src/shared/components/list/list-types/endpoint/endpoint-list.helpers.ts
@@ -14,7 +14,7 @@ import { SnackBarService } from '../../../../services/snackbar.service';
 import { ConfirmationDialogConfig } from '../../../confirmation-dialog.config';
 import { ConfirmationDialogService } from '../../../confirmation-dialog.service';
 import { createMetaCardMenuItemSeparator } from '../../../meta-card/meta-card-base/meta-card.component';
-import { IListAction } from '../../list.component.types';
+import { IListAction } from '../../../signal-list/list-action.types';
 import { TableCellCustom } from '../../list.types';
 
 interface EndpointDetailsContainerRefs {
