+++ src/frontend/packages/store/src/entity-catalog/entity-catalog.types.ts
@@ -3,7 +3,7 @@ import type { Store } from '@ngrx/store';
 import type { Observable } from 'rxjs';
 
 import type { HomePageEndpointCard } from '../../../core/src/features/home/home.types';
-import type { IListAction } from '../../../core/src/shared/components/list/list.component.types';
+import type { IListAction } from '../../../core/src/shared/components/signal-list/list-action.types';
 import type { AppState, GeneralEntityAppState } from '../app-state';
 import type {
   ApiErrorMessageHandler,
